+++ NEWS.md
@@ -8,6 +8,8 @@
   functions, if there is at least one missing value and `na.rm = FALSE` (the 
   default), then the output will be `NA`: `"max()"`, `"mean()"`, `"median()"`,
   `"min()"`, `"sd()"`, `"sum()"`, `"var()"` (#120).
+  
+* New argument `cluster_with_columns` in `collect()`, `compute()`, and `fetch()`.
 
 # tidypolars 0.8.0